Move `MainThreadMarker` to `objc2`
