+++ crates/objc2/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 * Added `AnyClass::is_metaclass`.
+* Added `MainThreadMarker` from `objc2-foundation`.
 
 
 ## 0.5.2 - 2024-05-21
